Consolidate librosa utils
