+++ examples/demos/public/src/index.js
@@ -12,7 +12,7 @@ export { computeSpectralCentroid, computeSpectrum, computeFFT } from './core/spe
 export { calculateBeatAlignment } from './core/musical-timing.js';
 
 // Basic audio processing
-export { loadAudioBuffer, computeRMS, computePeak, computeZeroCrossingRate } from './core/audio-utils.js';
+export { loadAudioBuffer, computeRMS, computePeak, computeZeroCrossingRate } from './utils/audio-utils.js';
 
 // Audio manipulation
 export { pitchBasedCompress, tempoBasedCompress } from './core/compression.js';
